django_project_904/requirements.txt

Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.2
django-cors-headers>=4.3
